broker-api refactor
