+++ BrokerApi/Configuration/RabbitSettings.cs
@@ -2,7 +2,7 @@
 
 namespace BrokerApi.Configuration;
 
-public class Rabbit
+public class RabbitSettings
 {
     public string? Url { get; set; }
     public string? UserName { get; set; }
